+++ scripts/rss-fetcher-stats.py
@@ -8,11 +8,11 @@ import logging
 import time
 from collections import Counter
 
-from sqlalchemy import func, select
+from sqlalchemy import Function, func, select
 
 from fetcher.config import conf
 from fetcher.database import Session
-from fetcher.database.models import Feed, Story
+from fetcher.database.models import Feed, Story, StoryRef
 from fetcher.logargparse import LogArgumentParser
 from fetcher.stats import Stats
 
@@ -121,14 +121,21 @@ def report_top_domain_stories(stats: Stats, days: int = TOP_SOURCE_DAYS) -> None
             break               # at most one row
 
 
-def report_stories_max_id(stats: Stats) -> None:
+def report_stories(stats: Stats) -> None:
     """
     report maximum Story.id to be able to show
     if/how-far story-indexer is behind
     """
+
     with Session() as session:
-        max_story = session.execute(select(func.max(Story.id))).scalar_one()
-    stats.gauge('stories.max-id', max_story)
+        def g(name: str, query: Function) -> None:
+            value = session.scalars(select(query)).one()
+            logger.debug("%s: %d", name, value)
+            stats.gauge(name, value)
+
+        g('stories.max-id', func.max(Story.id))
+        g('stories.count', func.count(Story.id))
+        g('story-refs.count', func.count(StoryRef.story_id))
 
 
 if __name__ == '__main__':
@@ -142,5 +149,5 @@ if __name__ == '__main__':
     while True:
         report_feeds_active(stats, 24)       # feeds active in last 24 hours
         report_top_domain_stories(stats)     # top domain story count
-        report_stories_max_id(stats)
+        report_stories(stats)
         time.sleep(args.interval - time.time() % args.interval)